split schema from plugin
