+++ tests/test_arcaflow_plugin_stressng.py
@@ -2,7 +2,7 @@
 
 import unittest
 import yaml
-import stressng_plugin
+import stressng_schema
 from arcaflow_plugin_sdk import plugin
 
 
@@ -10,20 +10,20 @@ class StressNGTest(unittest.TestCase):
     @staticmethod
     def test_serialization():
         plugin.test_object_serialization(
-            stressng_plugin.CpuStressorParams(stressor="cpu", cpu_count=2)
+            stressng_schema.CpuStressorParams(stressor="cpu", cpu_count=2)
         )
 
         plugin.test_object_serialization(
-            stressng_plugin.VmStressorParams(
+            stressng_schema.VmStressorParams(
                 stressor="vm", vm=2, vm_bytes="2G"
             )
         )
 
         plugin.test_object_serialization(
-            stressng_plugin.MatrixStressorParams(stressor="matrix", matrix=2)
+            stressng_schema.MatrixStressorParams(stressor="matrix", matrix=2)
         )
         plugin.test_object_serialization(
-            stressng_plugin.MqStressorParams(stressor="mq", mq=2)
+            stressng_schema.MqStressorParams(stressor="mq", mq=2)
         )
 
     def test_functional_cpu(self):
@@ -33,11 +33,11 @@ class StressNGTest(unittest.TestCase):
         # field names and can't do a direct
         # comparison of the returned values
 
-        cpu = stressng_plugin.CpuStressorParams(
+        cpu = stressng_schema.CpuStressorParams(
             stressor="cpu", cpu_count=2, cpu_method="all"
         )
 
-        stress = stressng_plugin.StressNGParams(
+        stress = stressng_schema.StressNGParams(
             timeout="10s", cleanup=True, items=[cpu]
         )
 
@@ -55,18 +55,18 @@ class StressNGTest(unittest.TestCase):
                 print(e)
 
         self.assertEqual(yaml.safe_load(result), reference)
-        workload_params = stressng_plugin.WorkloadParams(stress, True)
-        res = stressng_plugin.stressng_run(workload_params)
+        workload_params = stressng_schema.WorkloadParams(stress, True)
+        res = stressng_schema.stressng_run(workload_params)
         self.assertIn("success", res)
         self.assertEqual(res[1].cpuinfo.stressor, "cpu")
         self.assertGreaterEqual(res[1].cpuinfo.wall_clock_time, 10)
 
     def test_functional_vm(self):
-        vm = stressng_plugin.VmStressorParams(
+        vm = stressng_schema.VmStressorParams(
             stressor="vm", vm=1, vm_bytes="100m", mmap="1", mmap_bytes="10m"
         )
 
-        stress = stressng_plugin.StressNGParams(
+        stress = stressng_schema.StressNGParams(
             timeout="10s", cleanup=True, items=[vm]
         )
 
@@ -84,19 +84,19 @@ class StressNGTest(unittest.TestCase):
                 print(e)
 
         self.assertEqual(yaml.safe_load(result), reference)
-        workload_params = stressng_plugin.WorkloadParams(stress, True)
-        res = stressng_plugin.stressng_run(workload_params)
+        workload_params = stressng_schema.WorkloadParams(stress, True)
+        res = stressng_schema.stressng_run(workload_params)
         self.assertIn("success", res)
         self.assertEqual(res[1].vminfo.stressor, "vm")
         self.assertGreaterEqual(res[1].vminfo.wall_clock_time, 10)
 
     def test_functional_matrix(self):
-        matrix = stressng_plugin.MatrixStressorParams(
+        matrix = stressng_schema.MatrixStressorParams(
             stressor="matrix",
             matrix=1,
         )
 
-        stress = stressng_plugin.StressNGParams(
+        stress = stressng_schema.StressNGParams(
             timeout="10s", cleanup=True, items=[matrix]
         )
 
@@ -114,16 +114,16 @@ class StressNGTest(unittest.TestCase):
                 print(e)
 
         self.assertEqual(yaml.safe_load(result), reference)
-        workload_params = stressng_plugin.WorkloadParams(stress, True)
-        res = stressng_plugin.stressng_run(workload_params)
+        workload_params = stressng_schema.WorkloadParams(stress, True)
+        res = stressng_schema.stressng_run(workload_params)
         self.assertIn("success", res)
         self.assertEqual(res[1].matrixinfo.stressor, "matrix")
-        self.assertGreaterEqual(res[1].matrixinfo.wall_clock_time, 10)
+        self.assertGreaterEqual(res[1].matrixinfo.wall_clock_time, 9)
 
     def test_functional_mq(self):
-        mq = stressng_plugin.MqStressorParams(stressor="mq", mq=1)
+        mq = stressng_schema.MqStressorParams(stressor="mq", mq=1)
 
-        stress = stressng_plugin.StressNGParams(
+        stress = stressng_schema.StressNGParams(
             timeout="10s", cleanup=True, items=[mq]
         )
 
@@ -141,18 +141,18 @@ class StressNGTest(unittest.TestCase):
                 print(e)
 
         self.assertEqual(yaml.safe_load(result), reference)
-        workload_params = stressng_plugin.WorkloadParams(stress, True)
-        res = stressng_plugin.stressng_run(workload_params)
+        workload_params = stressng_schema.WorkloadParams(stress, True)
+        res = stressng_schema.stressng_run(workload_params)
         self.assertIn("success", res)
         self.assertEqual(res[1].mqinfo.stressor, "mq")
         self.assertGreaterEqual(res[1].mqinfo.wall_clock_time, 10)
 
     def test_functional_hdd(self):
-        hdd = stressng_plugin.HDDStressorParams(
+        hdd = stressng_schema.HDDStressorParams(
             stressor="hdd", hdd=1, hdd_bytes="100m", hdd_write_size="4m"
         )
 
-        stress = stressng_plugin.StressNGParams(
+        stress = stressng_schema.StressNGParams(
             timeout="10s", cleanup=True, items=[hdd]
         )
 
@@ -170,8 +170,8 @@ class StressNGTest(unittest.TestCase):
                 print(e)
 
         self.assertEqual(yaml.safe_load(result), reference)
-        workload_params = stressng_plugin.WorkloadParams(stress, True)
-        res = stressng_plugin.stressng_run(workload_params)
+        workload_params = stressng_schema.WorkloadParams(stress, True)
+        res = stressng_schema.stressng_run(workload_params)
         self.assertIn("success", res)
         self.assertEqual(res[1].hddinfo.stressor, "hdd")
         self.assertGreaterEqual(res[1].hddinfo.wall_clock_time, 10)
